Split DTO records into separate files and remove Lombok builder usage
