+++ backend/src/main/java/com/example/gymtracker/dto/CreateExerciseDTO.java
@@ -1,0 +1,5 @@
+package com.example.gymtracker.dto;
+
+import jakarta.validation.constraints.NotBlank;
+
+public record CreateExerciseDTO(@NotBlank String name, @NotBlank String muscleGroup) {}
