+++ backend/src/main/java/com/example/gymtracker/dto/CreateSessionDTO.java
@@ -1,0 +1,9 @@
+package com.example.gymtracker.dto;
+
+import jakarta.validation.constraints.NotNull;
+import java.time.LocalDate;
+import java.util.UUID;
+
+public record CreateSessionDTO(@NotNull LocalDate sessionDate,
+                               UUID workoutId,
+                               String notes) {}
