+++ backend/src/main/java/com/example/gymtracker/dto/ExerciseDTO.java
@@ -1,0 +1,5 @@
+package com.example.gymtracker.dto;
+
+import java.util.UUID;
+
+public record ExerciseDTO(UUID id, String name, String muscleGroup) {}
