+++ backend/src/main/java/com/example/gymtracker/web/SessionController.java
@@ -24,7 +24,7 @@ public class SessionController {
   public ResponseEntity<Map<String,Object>> get(@PathVariable UUID id){
     return sessions.findById(id)
       .map(s -> {
-        var payload = new HashMap<String,Object>();
+        Map<String,Object> payload = new HashMap<>();
         payload.put("session", s);
         payload.put("sets", sets.findBySessionIdOrderBySetIndexAsc(id));
         return ResponseEntity.ok(payload);
@@ -33,27 +33,29 @@ public class SessionController {
 
   @PostMapping
   public ResponseEntity<Session> create(@Valid @RequestBody CreateSessionDTO in){
-    Session s = Session.builder()
-      .userId(UUID.fromString("00000000-0000-0000-0000-000000000001"))
-      .workoutId(in.workoutId())
-      .sessionDate(in.sessionDate())
-      .notes(in.notes())
-      .build();
+    Session s = new Session(
+      null,
+      UUID.fromString("00000000-0000-0000-0000-000000000001"),
+      in.workoutId(),
+      in.sessionDate(),
+      in.notes()
+    );
     return ResponseEntity.status(HttpStatus.CREATED).body(sessions.save(s));
   }
 
   @PostMapping("/{id}/sets")
   public ResponseEntity<SetEntry> addSet(@PathVariable UUID id, @Valid @RequestBody CreateSetDTO in){
     if (!sessions.existsById(id)) return ResponseEntity.notFound().build();
-    SetEntry e = SetEntry.builder()
-      .sessionId(id)
-      .exerciseId(in.exerciseId())
-      .setIndex(in.setIndex())
-      .reps(in.reps())
-      .weightKg(in.weightKg())
-      .rpe(in.rpe())
-      .isPr(Boolean.TRUE.equals(in.isPr()))
-      .build();
+    SetEntry e = new SetEntry(
+      null,
+      id,
+      in.exerciseId(),
+      in.setIndex(),
+      in.reps(),
+      in.weightKg(),
+      in.rpe(),
+      Boolean.TRUE.equals(in.isPr())
+    );
     return ResponseEntity.status(HttpStatus.CREATED).body(sets.save(e));
   }
 }
